requirements.txt
oauthlib==3.1.0 
python-dotenv==0.12.0
requests-oauthlib==1.3.0 
tweepy==3.8.0